first channel one and zero of ADC1 was mixing the data fixed that, now for some reaseon the logic of the LM35 is inverted when i heat it up instead of go up it goes down idk why that happens i even replaced the analog sensor, and yeah tried 3 diffrent AI they cany solve shit

+++ Core/MyDRIVERS/ADC/MyAdcDMA.h
@@ -25,7 +25,7 @@ typedef enum{
 }adc_channels_t;
 typedef struct{
 	adc_channels_t channel[16];
-	uint8_t sample_time;
+	uint16_t sample_time;
 }adc_config_t;
 void adc_dma_init(ADC_TypeDef *adc_port,adc_config_t *ptr,DMA_TypeDef *DMAx,DMA_config_t* DMAx_config,uint8_t num_of_channels);
